+++ frontend/src/app/core/model/snippet.ts
@@ -4,7 +4,7 @@ export interface Snippet {
   codeSnippets: CodeSnippet[];
   tags: string[];
   userId?: string;
-  sourceUrl: string; // the location the codelet might have been inspired from to follow for further analysis
+  sourceUrl: string; // the location the snippet might have been inspired from to follow for further analysis
   public?: boolean;
   createdAt?: Date;
   updatedAt?: Date;